Allow removal of node if it does not contain regular subnodes
Once the assets definitions have been moved to the manifest file, the
module migrator attempts to remove the corresponding nodes. Previously,
it would not remove the nodes if it contained comments or processing
instructions (injected in order to preserve empty lines). Now, these
node types are ignored and the parent node is removed as intended.